+++ xarray_ms/backend/msv2/entrypoint_utils.py
@@ -38,7 +38,8 @@ def subtable_factory(
   try:
     return Table.from_filename(name, ninstances=1, readonly=True).to_arrow()
   except pa.lib.ArrowInvalid as e:
-    if "subtable" in e.msg and "is invalid" in e.msg:
+    e_str = str(e)
+    if "subtable" in e_str and "is invalid" in e_str:
       if on_missing == "raise":
         raise
       else: