+++ xarray_ms/backend/msv2/factories/field_and_source.py
@@ -13,6 +13,7 @@ class FieldAndSourceFactory(DatasetFactory):
   for a partition of the Measurement Set"""
 
   def get_dataset(self) -> Dataset:
+    import pyarrow as pa
     import pyarrow.compute as pac
 
     partition = self._structure_factory.instance[self._partition_key]
@@ -20,7 +21,8 @@ class FieldAndSourceFactory(DatasetFactory):
     source = self._subtable_factories["SOURCE"].instance
 
     field = maybe_impute_field_table(field, partition.field_ids)
-    source = maybe_impute_source_table(source, partition.source_ids)
+    source_ids = field["SOURCE_ID"].to_numpy()
+    source = maybe_impute_source_table(source, source_ids)
 
     num_poly = np.unique(field["NUM_POLY"].to_numpy())
     if not num_poly == [0]:
@@ -40,12 +42,22 @@ class FieldAndSourceFactory(DatasetFactory):
       )
 
     field_names = field["NAME"].to_numpy()
+    # Filter out negative source ids
+    pa_source_ids = pa.array(source_ids)
+    pa_source_ids = pac.replace_with_mask(
+      pa_source_ids,
+      pac.equal(pa_source_ids, -1),
+      pa.array([None] * len(source_ids), pa_source_ids.type),
+    )
+    pa_source_names = pac.take(source["NAME"], pa_source_ids)
+    source_names = pac.fill_null(pa_source_names, "UNKNOWN").to_numpy()
 
     return Dataset(
       data_vars=data_vars,
       coords={
         "field_name": Variable("field_name", field_names),
         "sky_dir_label": Variable("sky_dir_label", ["ra", "dec"]),
+        "source_name": Variable("Source_name", source_names),
       },
       attrs={"type": "field_and_source"},
     )