+++ doc/source/changelog.rst
@@ -5,8 +5,12 @@ Changelog
 
 X.Y.Z (DD-MM-YYYY)
 ------------------
+* Support TAI Measure encoding in time data (:pr:`124`)
+* Distinguish between time, baseline and channel irregularity (:pr:`123`)
+* Test xarray-ms against the full MSv4 Test Corpus (:pr:`120`)
+* Handle negative SOURCE_IDs in the FIELD subtable (:pr:`122`)
 * Cleanup Dataset Factories (:pr:`118`)
-* Add field_and_source_xds (:pr:`114`)
+* Add field_and_source_xds (:pr:`114`, :pr:`119`, :pr:`122`)
 * Fix integer selection (:pr:`117`)
 * Correct clamp of the slice stop to be min(shape[d], s.stop) (:pr:`115`)
 * Rename deprecated ``repository_url`` to ``repository-url`` (:pr:`112`)